[Tutorials] Support base64 image attachments
The old tutorials were including images using base64 attachments which were not properly supported by the conversion script.

These image attachments are stored in the notebook cell's "attachments" field in base64 format with their associated mime_type and referenced in the markdown via attachment name.

The pattern we search for in the Markdown is `![alt_text](attachment:attachment_name title)` with three groups:

- group 1 = alt_text (optional)
- group 2 = attachment_name
- group 3 = title (optional)

To represent this in MD we replace the attachment reference with the base64 encoded string as `![{alt_text}](data:{mime_type};base64,{img_as_base64})`

This fix won't automatically propogate to the broken old tutorials, in a separate commit I'll fix them by updating the pre-built tutorial mdx stored in the `docusaurus-versions` branch